split user accounts model tests into separate files
